+++ src/main/java/net/frozenblock/wilderwild/mixin/block/echo_glass/client/MultiPlayerGameModeMixin.java
@@ -16,7 +16,7 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.mixin.snowlogging.client;
+package net.frozenblock.wilderwild.mixin.block.echo_glass.client;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
@@ -25,17 +25,22 @@ import com.llamalad7.mixinextras.sugar.Share;
 import com.llamalad7.mixinextras.sugar.ref.LocalRef;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
+import net.frozenblock.wilderwild.block.EchoGlassBlock;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 
 @Environment(EnvType.CLIENT)
 @Mixin(MultiPlayerGameMode.class)
-public class MultiPlayerGameModeMixin {
+public abstract class MultiPlayerGameModeMixin {
+
+	@Shadow
+	public abstract GameType getPlayerMode();
 
 	@ModifyExpressionValue(
 		method = "destroyBlock",
@@ -44,7 +49,7 @@ public class MultiPlayerGameModeMixin {
 			target = "Lnet/minecraft/world/level/block/Block;playerWillDestroy(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/entity/player/Player;)Lnet/minecraft/world/level/block/state/BlockState;"
 		)
 	)
-	public BlockState wilderWild$destroyBlockA(
+	public BlockState wilderWild$destroyBlockEchoA(
 		BlockState original,
 		@Share("wilderWild$destroyedState") LocalRef<BlockState> destroyedState
 	) {
@@ -59,12 +64,13 @@ public class MultiPlayerGameModeMixin {
 			target = "Lnet/minecraft/world/level/Level;setBlock(Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;I)Z"
 		)
 	)
-	public boolean wilderWild$destroyBlockB(
+	public boolean wilderWild$destroyBlockEchoB(
 		Level instance, BlockPos pos, BlockState newState, int flags, Operation<Boolean> original,
 		@Share("wilderWild$destroyedState") LocalRef<BlockState> destroyedState
 	) {
-		if (SnowloggingUtils.isSnowlogged(destroyedState.get())) {
-			instance.setBlock(pos, destroyedState.get().setValue(SnowloggingUtils.SNOW_LAYERS, 0), flags);
+		BlockState blockState = destroyedState.get();
+		if (blockState.getBlock() instanceof EchoGlassBlock && EchoGlassBlock.canDamage(blockState) && !this.getPlayerMode().isCreative()) {
+			EchoGlassBlock.setDamagedState(instance, pos, blockState);
 			return true;
 		}
 		return original.call(instance, pos, newState, flags);